MUSEBOARD/requirements.txt
flask
pillow
python-dotenv
groq
deep-translator
gtts 
arabic_reshaper